chore(rebase): implement git rebase to resolve conflictensure all conflict were resolved[Delivers #165570919]
