+++ package.json
@@ -36,7 +36,8 @@
     "express": "^4.16.4",
     "express-validator": "^5.3.1",
     "jsonwebtoken": "^8.5.1",
-    "pg": "^7.9.0"
+    "pg": "^7.9.0",
+    "swagger-ui-express": "^4.0.2"
   },
   "devDependencies": {
     "coveralls": "^3.0.3",
