--- server_db/middlewares/validateUser.js
@@ -1,145 +1,0 @@
-/* eslint-disable consistent-return */
-class ValidateUser {
-  static validateSignup(request, response, next) {
-    const {
-      firstName, lastName, email, password, phoneNumber,
-    } = request.body;
-    const pattern = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-    if (!firstName) {
-      return response.status(400).json({
-        status: 400,
-        error: 'firstname is required',
-      });
-    }
-    if (!firstName.match(/[^\s-]/)) {
-      return response.status(400).json({
-        status: 400,
-        error: 'Spaces are not allowed',
-      });
-    }
-    if (!isNaN(firstName)) {
-      return response.status(400).json({
-        status: 400,
-        error: 'Firstname must be letters',
-      });
-    }
-    if (firstName.length < 2) {
-      return response.status(400).json({
-        status: 400,
-        error: 'first Name must be atleast 3 alphabets',
-      });
-    }
-    if (!lastName) {
-      return response.status(400).json({
-        status: 400,
-        error: 'lastname is required',
-      });
-    }
-    if (!lastName.match(/[^\s-]/)) {
-      return response.status(400).json({
-        status: 400,
-        error: 'Spaces are not allowed',
-      });
-    }
-    if (!isNaN(lastName)) {
-      return response.status(400).json({
-        status: 400,
-        error: 'Lastname must be letters',
-      });
-    }
-    if (lastName.length < 2) {
-      return response.status(400).json({
-        status: 400,
-        error: 'last Name must be atleast 3 alphabets',
-      });
-    }
-    if (!email) {
-      return response.status(400).json({
-        status: 400,
-        error: 'email is required',
-      });
-    }
-    if (!email.match(/[^\s-]/)) {
-      return response.status(400).json({
-        status: 400,
-        error: 'Spaces are not allowed',
-      });
-    }
-    if (!email.match(pattern)) {
-      return response.status(400).json({
-        status: 400,
-        error: 'Please provide a valid email',
-      });
-    }
-    if (!password) {
-      return response.status(400).json({
-        status: 400,
-        error: 'password is required',
-      });
-    }
-
-    if (!phoneNumber) {
-      return response.status(400).json({
-        status: 400,
-        error: 'phoneNumber is required',
-      });
-    }
-
-    if (phoneNumber.toString().replace(/\s/g, '').length === 0) {
-      return response.status(400).json({
-        status: 400,
-        error: 'Phone Number must not empty or white-space ',
-      });
-    }
-  
-    /* Check if phhone is a number */
-    if (isNaN(phoneNumber)) {
-      return response.status(400).json({
-        status: 400,
-        error: 'Phone Number must be a number ',
-      });
-    }
-  
-    /* Check if phone is a whole number */
-    if ((phoneNumber % 1) !== 0) {
-      return response.status(400).json({
-        status: 400,
-        error: 'Phone Number must be a positive integer ',
-      });
-    }
-    next();
-  }
-
-  static validateLogin(request, response, next) {
-    const { email, password } = request.body;
-    const pattern = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-    if (!email) {
-      return response.status(400).json({
-        status: 400,
-        error: 'email is required',
-      });
-    }
-    if (!email.match(/[^\s-]/)) {
-      return response.status(400).json({
-        status: 400,
-        error: 'Spaces are not allowed',
-      });
-    }
-    if (!email.match(pattern)) {
-      return response.status(400).json({
-        status: 400,
-        error: 'Please provide a valid email',
-      });
-    }
-    if (!password) {
-      return response.status(400).json({
-        status: 400,
-        error: 'password is required to login',
-      });
-    }
-    next();
-  }
-}
-
-
-export default ValidateUser;
