+++ server_db/routes/accountRoute.js
@@ -1,27 +1,22 @@
 import express from 'express';
 import accountController from '../controllers/accountController';
-import auth from '../middlewares/authentication';
-import role from '../middlewares/authorization';
-import ValidateAmount from '../middlewares/validateAmount';
-import ValidateAccType from '../middlewares/validateBankAccount';
-import validateAccountNumber from '../middlewares/validateAccountNumber';
-import ValidateEmail from '../middlewares/validateEmail';
-import ValidateStatus from '../middlewares/validateStatus';
+import authenticate from '../middlewares/authentication';
+import Validate from '../middlewares/Validate';
 
 const router = express.Router();
 
-router.post('/accounts', auth, ValidateAccType.validateType, ValidateAmount.validateAmt, accountController.postAccount);
+router.post('/accounts', authenticate, Validate.validateAccountType, Validate.validateAmount, accountController.postAccount);
 
-router.patch('/account/:accountNumber', auth, validateAccountNumber, ValidateStatus.validateStatus,
-  role.admin, accountController.editAccountStatus);
+router.patch('/account/:accountNumber', authenticate, Validate.admin, Validate.validateAccountNumber, Validate.validateAccountStatus,
+  accountController.editAccountStatus);
 
-router.delete('/accounts/:accountNumber', auth, validateAccountNumber,
-  role.admin, accountController.deleteAccount);
+router.delete('/accounts/:accountNumber', authenticate, Validate.admin, Validate.validateAccountNumber,
+  accountController.deleteAccount);
 
-router.get('/accounts/:accountNumber', auth, validateAccountNumber, accountController.getByAccountNumber);
+router.get('/accounts/:accountNumber', authenticate, Validate.validateAccountNumber, accountController.getByAccountNumber);
 
-router.get('/user/:email/accounts', auth, ValidateEmail, role.admin, accountController.getAccountByEmail);
+router.get('/user/:email/accounts', authenticate, Validate.validateEmail, accountController.getAccountByEmail);
 
-router.get('/accounts', auth, role.admin, accountController.getAllAccounts);
+router.get('/accounts', authenticate, Validate.admin, accountController.getAllAccounts);
 
 export default router;
