+++ server_db/routes/userRoute.js
@@ -1,10 +1,10 @@
 import express from 'express';
 import UserController from '../controllers/userController';
-import ValidateUser from '../middlewares/validateUser';
+import Validate from '../middlewares/Validate';
 
 const router = express.Router();
 
-router.post('/auth/signup', ValidateUser.validateSignup, UserController.userSignup);
-router.post('/auth/signin', ValidateUser.validateLogin, UserController.userLogin);
+router.post('/auth/signup', Validate.validateEmail, Validate.validateSignup, UserController.userSignup);
+router.post('/auth/signin', Validate.validateEmail, Validate.validatePassword, UserController.userLogin);
 
 export default router;
