+++ server_db/tests/userTest.js
@@ -182,4 +182,3 @@ describe('/POST Login user', () => {
       });
   });
 });
-
